pytest/requirements.txt
selenium==4.3.0
selenium-wire==4.6.5
Pillow==9.2.0